functions/pubsub-receiver/requirements.txt
google-cloud-pubsub==2.13.0
flask==2.0.1
werkzeug==2.0.1
functions-framework==3.0.0
google-cloud-language==2.4.0
requests==2.28.0
google-auth==2.15.0